fix(audio): make system stop results target-neutral
